Reduce overhead in ctypes.string_at() when grabbing from a region smaller than the whole screen.
The idea is to ask ctypes.string_at for as little memory as possible. Since images are stored in memory with width being the fast index. If we want to grab a 480x640 region from a 1440x2560 screen we can ask ctypes.string_at() for a 480x2560 region. This reduces memory allocation and memcpy overhead in ctypes.string_at().
To grab a 480x640 region out of a 1440x2560 screen the profiler time spent went from ~24% to ~8%.